refactor: move add_cooperative_yielding into leveled_store
# Summary

The stream-yielding helper moves from the crate-level `utils` into
`leveled_store::util`. `leveled_store` now has no intra-crate dependencies at
all.

# Details

`utils` was a grab bag of three unrelated helpers with different audiences.
`add_cooperative_yielding` is a stream adapter used by the storage engine and
by the two KV read paths above it, while `prefix_right_bound` and
`seq_marked_to_seqv` are key-range and value-conversion helpers that only the
state machine and the service use. Keeping them together forced the engine to
depend on a module it shares with its own callers.

Splitting along that line leaves `leveled_store` self-contained, which is a
precondition for extracting it as a crate that depends on nothing else of
ours. The helper becomes `pub` since it is used from outside the storage
engine.
