+++ crates/server/raft-store/src/state_machine_api_ext.rs
@@ -47,7 +47,7 @@ use state_machine_api::StateMachineApi;
 use state_machine_api::UserKey;
 
 use crate::leveled_store::map_api::MapApiHelper;
-use crate::utils::add_cooperative_yielding;
+use crate::leveled_store::util::add_cooperative_yielding;
 use crate::utils::prefix_right_bound;
 use crate::utils::seq_marked_to_seqv;
